moved test methods
